Move Enclosure ADT + enclosingScope from a separate Enclosures trait into Environments (env: MacroCommons self-type preserves the env alias; qualify 4 java.lang.Class refs shadowed by the cake's Class[A]); delete Enclosures.scala. The Scala 3 Expr/Position cycle that forced the split does not recur — it was an artifact of the separate trait, not the placement
